backend/requirements.txt
Flask==2.1.1
Flask-Cors==3.0.10
PyJWT==2.3.0
requests==2.26.0
python-dotenv==0.19.2
gunicorn==20.1.0
python-multipart==0.0.5
pydantic>=2.0.0
pytest==6.2.5
pytest-cov==2.12.1
Werkzeug<2.1.0